+++ eras/conway/impl/CHANGELOG.md
@@ -13,6 +13,10 @@
 
 * Add `DecCBOR` instance for `Block`
 
+### `cddl`
+
+* Add `HuddleRule "vrf_cert"` instance
+
 ## 1.23.0.0
 
 * Add `Ord` instances for `ConwayBbodyPredFailure`, `ConwayCertPredFailure`,